+++ src/pyclass_init.rs
@@ -1,6 +1,6 @@
 //! Contains initialization utilities for `#[pyclass]`.
 use crate::class::impl_::PyClassThreadChecker;
-use crate::pyclass_slots::{PyClassDict, PyClassWeakRef};
+use crate::impl_::pyclass::{PyClassDict, PyClassWeakRef};
 use crate::{callback::IntoPyCallbackOutput, class::impl_::PyClassBaseType};
 use crate::{ffi, PyCell, PyClass, PyErr, PyResult, Python};
 use crate::{